Move contentTypeChecks
