+++ packages/react-openapi/src/contentTypeChecks.ts
